requirements.txt
# NOTICE: This file is automatically generated by scripts/export_requirements.py
# This is also automatically regenerated when an edit to pyproject.toml or poetry.lock is commited.

aiodns==3.0.0
aiohttp==3.7.4.post0 ; python_version >= "3.6"
arrow==1.1.1 ; python_version >= "3.6"
async-timeout==3.0.1 ; python_full_version >= "3.5.3"
atoml==1.0.3 ; python_version >= "3.6"
attrs==21.2.0 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2" and python_version != "3.3" and python_version != "3.4"
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.15.0
chardet==4.0.0 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2" and python_version != "3.3" and python_version != "3.4"
colorama==0.4.4 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2" and python_version != "3.3" and python_version != "3.4"
coloredlogs==15.0.1 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2" and python_version != "3.3" and python_version != "3.4"
desert==2020.11.18 ; python_version >= "3.6"
discord.py @ https://github.com/Rapptz/discord.py/archive/45d498c1b76deaf3b394d17ccf56112fa691d160.zip ; python_full_version >= "3.8.0"
humanfriendly==10.0 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2" and python_version != "3.3" and python_version != "3.4"
idna==3.2 ; python_version >= "3.5"
marshmallow-enum==1.5.1
marshmallow==3.13.0 ; python_version >= "3.5"
multidict==5.2.0 ; python_version >= "3.6"
mypy-extensions==0.4.3
pycares==4.1.2
pycparser==2.20 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2" and python_version != "3.3"
pyreadline3==3.3 ; sys_platform == "win32"
python-dateutil==2.8.2 ; python_version != "3.0"
python-dotenv==0.19.2 ; python_version >= "3.5"
pyyaml==5.4.1 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2" and python_version != "3.3" and python_version != "3.4" and python_version != "3.5"
six==1.16.0 ; python_version >= "2.7" and python_version != "3.0" and python_version != "3.1" and python_version != "3.2"
typing-extensions==3.10.0.2
typing-inspect==0.7.1
yarl==1.7.2 ; python_version >= "3.6"
